+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `[options]` Added method `Delete`
 - `[usage]` Added support for optional arguments to usage info
 - `[uuid]` Added [UUID7](https://uuid7.com) generator
+- `[options]` Code refactoring
 
 ### [12.126.1](https://kaos.sh/ek/12.126.1)